:recycle: DontForgetMashUpView파일 분리
